+++ src/app/api/menu-items/route.js
@@ -1,25 +1,34 @@
 import mongoose from "mongoose"
 import { MenuItem } from "../../models/MenuItem"
+import { isAdmin } from "../auth/[...nextauth]/route"
 
 export async function POST(req) {
 mongoose.connect(process.env.MONGO_URL)
 const data = await req.json()
-const menuItemDoc = await MenuItem.create(data)
-return Response.json(menuItemDoc)
+if(await isAdmin()){
+    const menuItemDoc = await MenuItem.create(data)
+    return Response.json(menuItemDoc)
+}
 }
 
 //get
 export async function GET() {
     mongoose.connect(process.env.MONGO_URL)
-    const menuItems = await MenuItem.find()
-    return Response.json(menuItems)
+    if(await isAdmin()){
+        const menuItems = await MenuItem.find()
+        return Response.json(menuItems)
+    }else{
+        return Response.json([])
+    }
 }
 
 //put
 export async function PUT(req) {
     mongoose.connect(process.env.MONGO_URL)
-    const {_id, ...data} = await req.json()
-    await MenuItem.findByIdAndUpdate(_id, data)
+    if(await isAdmin()){
+        const {_id, ...data} = await req.json()
+        await MenuItem.findByIdAndUpdate(_id, data)
+    }
     return Response.json(true)
 }
 
@@ -28,6 +37,8 @@ export async function DELETE(req) {
     mongoose.connect(process.env.MONGO_URL)
     const url = new URL(req.url)
     const _id = url.searchParams.get('_id')
-    await MenuItem.deleteOne({_id})
+    if(await isAdmin()){
+        await MenuItem.deleteOne({_id})
+    }
     return Response.json(true)
 }
